updatecarousel

+++ controllers/adminController.js
@@ -8,7 +8,7 @@ const fs = require('fs');
 const path = require('path');
 const Calender = require('../models/Calender');
 const crms = require('../models/crmModel');
-
+const assignment = require('../models/Assignments');
 // register
 // const register = async (req, res) => {
 //     console.log("Inside register API");
@@ -170,6 +170,9 @@ const updateCarousel = async (req, res) => {
         const { id } = req.params;
         const { title, description, link } = req.body;
         const imgObj = req.file;
+        if(!imgObj || !imgObj.filename){
+            return res.status(400).json({ error: "Image file is missing or invalid" });
+        }
 
         const updatedCarousel = await Carousel.findByIdAndUpdate(id, {
             title,
@@ -181,18 +184,16 @@ const updateCarousel = async (req, res) => {
             return res.status(404).json({ error: "Carousel not found" });
         }
 
-
         // Construct the path to the old image file
         const oldImageFilename = updatedCarousel.image;
         const filename = oldImageFilename.split('/').pop(); // Get the last part (filename)
         const oldImagePath = path.join('public', 'carousel', filename);
 
-        // Construct the path to the new image file
-        const newImageFilename = imgObj.filename;
-        const newImagePath = path.join('public', 'carousel', newImageFilename);
-
-        // Delete the old image file
-        fs.unlinkSync(oldImagePath);
+        // Check if the old image file exists before trying to delete it
+        if (fs.existsSync(oldImagePath)) {
+            // Delete the old image file
+            fs.unlinkSync(oldImagePath);
+        }
 
         // Update the carousel cache
         const carouselCache = await Carousel.find().sort({ _id: -1 });
@@ -357,6 +358,28 @@ const updateCrm =async (req,res)=>{
     }
 };
 
+const addAssignments = async (req, res) => {
+    try {
+        const { title, subject, assignmentType, issueDate, dueDate, priority,createdBy } = req.body;
+        const assignments = await assignment.create({
+            title,
+            subject,
+            assignmentType,
+            issueDate,
+            dueDate,
+            priority,
+            createdBy: req.admin.id
+            
+        });
+        const cacheDate = await assignment.find().sort({ _id: -1 });
+        Cache.set('assignment', cacheDate, catchTime);
+        res.status(200).json({ assignments });
+    } catch (error) {
+        res.status(500).json({ error: "Internal Server Error", message: error.message });
+    }   
+};
+
+
 module.exports = {
     // register,
     adminLogin,
@@ -372,4 +395,5 @@ module.exports = {
     getCarouselById,
     getCrm,
     updateCrm,
+    addAssignments
 }
